+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/design-tokens
 
+## 0.0.44
+
+### Patch Changes
+
+- update dark style
+
 ## 0.0.43
 
 ### Patch Changes